requirements.txt
django-csp
sentry-sdk
safety
bandit
